Abstract out target setting code
